Add support for pre-1.0 Tiled JSON formats
While attempting to parse Starbound dungeons in Tiled JSON format, I
encountered a number of parse errors. These errors appear to be due to
pre-1.0 differences in the Tiled JSON map and tileset formats. As far as
I can tell, most Starbound dungeon files were created and/or last edited
with Tiled 0.15.2.

* Map object `properties` were dictionaries rather than lists, with the
  key being the property name and the value always a string.
* Tileset `tiles` were dictionaries rather than lists, with the key
  being the tile id (integer cast as a string) and the value being a
  dictionary containing the remainder of the tile definition.
* Some map and tileset properties did not exist, or were optional.

This patch introduces minimal changes to the code to allow the Starbound
files to be parsed. There may be other legacy quirks, but these are the
only ones I've noticed in these files.